Separate MapMarker into new component file
